requirements.txt
gradio
google-generativeai
python-dotenv
Pillow
json
time
openai
urllib3
certifi
charset-normalizer
idna
python-dateutil
pytz
requests
six
urllib3
websockets




